requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.1
psycopg2-binary==2.9.9
sqlparse==0.5.3
tzdata==2025.2
